course work/kinovik/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.27.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-filter==21.1
django-heroku==0.3.1
django-taggit==1.5.1
fontawesome-free==5.15.4
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
psycopg2==2.9.2
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.27.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-taggit==1.5.1
fontawesome-free==5.15.4
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
psycopg2==2.9.2
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.27.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-taggit==1.5.1
fontawesome-free==5.15.4
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
psycopg2==2.9.2
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.27.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-taggit==1.5.1
fontawesome-free==5.15.4
future==0.18.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.27.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.5.1
fontawesome-free==5.15.4
future==0.18.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
